+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## master
+##### Bugfix & Breaking
+* [fixed] Quoted parameter is unquoted [#49](https://github.com/toshi0383/cmdshelf/issues/49)  
+  [Toshihiro Suzuki](https://github.com/toshi0383)
+  [#50](https://github.com/toshi0383/cmdshelf/issues/50)  
+  **Breaking**: This changes the way subcommand is evaluated. Double/Single quoted alias parameter argument does not work any more.
+
 ## 0.8.0
 ##### Enhancements
 * No need to quote command with parameters